third_party/blink/web_tests/fast/forms/textarea/textarea-newline-expected.txt

These tests check if leading line feeds characters in textarea as default values are parsed correctly.
    

PASS document.getElementById("no-line-feed").value is "Madoka"
PASS document.getElementById("one-line-feed").value is "Sayaka"
PASS document.getElementById("two-line-feeds").value is "\nMami"
PASS document.getElementById("one-line-feed-escaped-char-and-one-line-feed").value is "\nKyoko"
PASS document.getElementById("two-line-feed-escaped-chars").value is "\nHomura"
PASS successfullyParsed is true

TEST COMPLETE
